+++ index.js
@@ -1,1 +1,7 @@
-module.exports = require('./lib');
+module.exports = {
+  generate: require('./lib'),
+  parsers: {
+    javascript: require('./lib/parse/javascript'),
+    typescript: require('./lib/parse/typescript')
+  }
+};